refactor(ui): moved useDnd & useBatch to global hooks
