+++ ui/src/features/Editor/components/Canvas/useNodes.ts
@@ -12,10 +12,11 @@ import {
 } from "@xyflow/react";
 import { MouseEvent, useCallback } from "react";
 
+import useBatch from "@flow/hooks/useBatch";
+import useDnd from "@flow/hooks/useDnd";
 import type { ActionNodeType, Edge, Node } from "@flow/types";
 
-import useBatch from "./useBatch";
-import useDnd from "./useDnd";
+// import useDnd from "./useDnd";
 
 type Props = {
   nodes: Node[];
